// src/types/indy.ts
export enum IndyTypes {
    BP_MANU = "BP_MANU",
    ORDER_FULFILLMENT = "ORDER_FULFILLMENT",
    RIG_MANU = "RIG_MANU",
    SHIP_MANU_BATTLESHIP = "SHIP_MANU_BATTLESHIP",
    SHIP_MANU_BCRUISER = "SHIP_MANU_BCRUISER",
    SHIP_MANU_CAPITAL = "SHIP_MANU_CAPITAL",
    SHIP_MANU_CRUISER = "SHIP_MANU_CRUISER",
    SHIP_MANU_DESTROYER = "SHIP_MANU_DESTROYER",
    SHIP_MANU_FRIGATE = "SHIP_MANU_FRIGATE",
    STRUCTURE_MANU = "STRUCTURE_MANU",
}